rename channel card
