--- src/shared/components/ChannelCard/index.ts
@@ -1,3 +1,0 @@
-import ChannelCard from './ChannelCard'
-
-export default ChannelCard
