+++ src/shared/components/StudioCard/StudioCard.stories.tsx
@@ -1,10 +1,10 @@
 import { Meta, Story } from '@storybook/react'
 import React from 'react'
-import ChannelCard, { ChannelCardProps } from './ChannelCard'
+import StudioCard, { StudioCardProps } from './StudioCard'
 
 export default {
-  title: 'Shared/ChannelCard',
-  component: ChannelCard,
+  title: 'Shared/StudioCard',
+  component: StudioCard,
   argTypes: {
     avatarPhotoUrl: {
       defaultValue: 'https://eu-central-1.linodeobjects.com/atlas-assets/channel-posters/2.jpg',
@@ -12,6 +12,6 @@ export default {
   },
 } as Meta
 
-const Template: Story<ChannelCardProps> = (args) => <ChannelCard {...args} />
+const Template: Story<StudioCardProps> = (args) => <StudioCard {...args} />
 
 export const Default = Template.bind({})
