+++ src/shared/components/StudioCard/StudioCard.style.ts
