+++ src/shared/components/StudioCard/index.ts
@@ -1,0 +1,3 @@
+import StudioCard from './StudioCard'
+
+export default StudioCard
