+++ src/shared/components/index.ts
@@ -38,7 +38,7 @@ export { default as FileSelect } from './FileSelect'
 export { default as FileStep } from './FileStep'
 export { default as TopbarBase, TOP_NAVBAR_HEIGHT } from './TopbarBase'
 export type { TopbarBaseProps } from './TopbarBase'
-export { default as ChannelCard } from './ChannelCard'
+export { default as StudioCard } from './StudioCard'
 export { default as StudioHeader } from './StudioHeader'
 export { default as Spinner } from './Spinner'
 export { default as AccountBar } from './AccountBar'
